% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mi_stats.R
\name{check_original_data}
\alias{check_original_data}
\title{Internal function for checking presence of the "missing data"}
\usage{
check_original_data(incomplete_data, full_data, ids = ".id",
  imps = ".imp")
}
\arguments{
\item{incomplete_data, full_data}{datasets with incomplete and full data for comparison}

\item{ids}{names of the columns wit the observation-specific ID and
number of imputation.}

\item{imps}{names of the columns wit the observation-specific ID and
number of imputation.}
}
\description{
Internal function for checking presence of the "missing data"
}
