% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bootstrap_permutation.R
\name{convert_bootstrap_permutation}
\alias{convert_bootstrap_permutation}
\title{Converts a table of the values into a list of permutation vectors. (used for development)}
\usage{
convert_bootstrap_permutation(boot_table)
}
\arguments{
\item{boot_table}{is the table of the bootstrap iterations.}
}
\description{
Converts a table of the values into a list of permutation vectors. (used for development)
}
