% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_near.R
\name{find_near}
\alias{find_near}
\title{Find neares match between a vector of observations and a vector of alternatives}
\usage{
find_near(observ_vector, match_vector, n_near = 5)
}
\arguments{
\item{observ_vector}{named vector or value to find a match from `match_vector`.
If the vector has no names function will still work, but the resulting
`target_id` variable will be empty.}

\item{match_vector}{named vector which is used to find a nearest match to the
`observ_vector`.}

\item{n_near}{number of the nearest observations to derive a random match.
If `n_near` is greater than `length(match_vector)`, minimum out of two is
used to create a sample for selecting a random match value.}
}
\value{
A dataframe with the number of rows equall to the lenght of the
  `observ_vector` vector and one mathcing value and its index from the
  `match_vector` corresponding to each value afrom the `observ_vector`.
}
\description{
Find neares match between a vector of observations and a vector of alternatives
}
\examples{
library(dplyr)
library(purrr)
library(glmnet)
library(syntheticpanel)

obs <- setNames(rnorm(5), as.character(1:5))
match_v <- setNames(rnorm(50), as.character(1:50))
find_near(obs, match_v, n_near = 10)
}
