% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_combinations.R
\name{get_all_combinations}
\alias{get_all_combinations}
\title{Convert a table of dummy variables to a table of its unique combinations}
\usage{
get_all_combinations(.data, new_var, clean_name = TRUE,
  clean_expr = "_.{1,}$", connector = "_")
}
\arguments{
\item{.data}{data frame with only dummy variable to be converted}

\item{new_var}{name of the new categorical variable to create for
representing all combinations of the previous dummy variables.}
}
\description{
Convert a table of dummy variables to a table of its unique combinations
}
\examples{
library(dplyr)
library(tibble)
tibble(
       a = sample(c(0L, 1L), 25, replace = TRUE),
       b = sample(c(0L, 1L), 25, replace = TRUE)
       ) \%>\%
  get_all_combinations(c)


}
