% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mi_stats.R
\name{get_mi_mean}
\alias{get_mi_mean}
\title{Get multiple imputation mean statistics}
\usage{
get_mi_mean(.data, stat_var, weight_var = NULL, ids = ".id",
  imps = ".imp", summary = TRUE, use_random = FALSE)
}
\arguments{
\item{.data}{dataframe that could be converted to the mids object to perform
calculations on the multiple imputations data frame.}

\item{stat_var}{variable to apply statistics}

\item{weight_var}{variable for weighting. NULL by default.}

\item{ids, imps}{names of the columns wit the observation-specific ID and
number of imputation.}

\item{summary}{return a summary of the pooled analysis.}

\item{use_random}{harmonise imputation of the unequal size randomly. By
default if FALSE, therefore, to equalise sizes of the imputed samples,
simplu n-first observations is used.}
}
\description{
Get multiple imputation mean statistics
}
\examples{
library(dplyr)
dep <- "price"            # Dependent variable
indep <- c("mpg", "headroom", "trunk",
           "weight", "length", "turn",
           "displacement", "gear_ratio",
           "foreign")     # independent variables
weight <- "weight"        # weight variable
extrar <- "displacement"  # any extra variable to bring from period 0 data
bt_groups <- c("psu")     # Grouping variable for bootsrapping.
n_nearest <- 5    # numebr of the nearest observations to drow a random match
set.seed(11344)
imputation <-
  lassopmm(
    source = p_0_exmple,
    target = p_1_exmple,
    dep_var = dep,
    indep_var = indep,
    weight_var = weight,
    strata_vars = bt_groups,
    extra_var = extrar,
    n_boot = 5, # Numebr of bootstrap iterations
    n_near = 1)

# Getting results of the averages
get_mi_mean(imputation, stat_var = "price_source", use_random = F)

# Same by with weights in means
get_mi_mean(imputation, stat_var = "price_source", weight_var = "weight")

}
