% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mi_stats.R
\name{get_mi_mean_by_group}
\alias{get_mi_mean_by_group}
\title{Get multiple imputation mean statistics using gouping variable.}
\usage{
get_mi_mean_by_group(full_data, stat_var, group_var, weight_var = NULL,
  ids = ".id", imps = ".imp", use_random = FALSE)
}
\arguments{
\item{full_data}{ataframe that could be converted to the mids object to perform
calculations on the multiple imputations data frame.}

\item{stat_var}{variable to apply statistics}

\item{group_var}{variable for grupping}

\item{weight_var}{variable for weighting. NULL by default.}

\item{ids}{names of the columns wit the observation-specific ID and
number of imputation.}

\item{imps}{names of the columns wit the observation-specific ID and
number of imputation.}

\item{use_random}{harmonise imputation of the unequal size randomly. By
default if FALSE, therefore, to equalise sizes of the imputed samples,
simplu n-first observations is used.}
}
\description{
Get multiple imputation mean statistics using gouping variable.
}
