% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mi_stats.R
\name{get_mi_means_table}
\alias{get_mi_means_table}
\title{Get multiple imputation mean statistics for several variables}
\usage{
get_mi_means_table(.data, stat_variables_vec, weight_var = NULL,
  ids = ".id", imps = ".imp", use_random = FALSE)
}
\arguments{
\item{.data}{dataframe that could be converted to the mids object to perform
calculations on the multiple imputations data frame.}

\item{stat_variables_vec}{is the atomic vector with variables names}

\item{weight_var}{variable for weighting. NULL by default.}

\item{ids}{names of the columns wit the observation-specific ID and
number of imputation.}

\item{imps}{names of the columns wit the observation-specific ID and
number of imputation.}

\item{use_random}{harmonise imputation of the unequal size randomly. By
default if FALSE, therefore, to equalise sizes of the imputed samples,
simplu n-first observations is used.}
}
\description{
Get multiple imputation mean statistics for several variables
}
