% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_0_exmple}
\alias{p_0_exmple}
\title{Sample dataset on automobiles borrowed from STATA.}
\format{A data frame with 74 rows and 16 variables}
\source{
\url{http://www.stata-press.com/data/r15/auto.dta}
}
\usage{
p_0_exmple
}
\description{
Sample dataset on automobiles borrowed from STATA.
}
\keyword{datasets}
