% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_1_exmple}
\alias{p_1_exmple}
\title{Sub-sample of the automobiles data from STATA.}
\format{A data frame with 7 rows and 16 variables}
\source{
\url{http://www.stata-press.com/data/r15/auto.dta}
}
\usage{
p_1_exmple
}
\description{
Sub-sample of the automobiles data from STATA.
}
\keyword{datasets}
