% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny_example.R
\name{run_shiny_app_name}
\alias{run_shiny_app_name}
\title{Run shiny app available in the package}
\usage{
run_shiny_app_name(app_name = "shiny_imputations")
}
\arguments{
\item{app_name}{is the name of the app availabe in the package.}
}
\description{
Run shiny app available in the package
}
