% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_bootstrap}
\alias{sample_bootstrap}
\title{Bootstrap permutation vectors sample dataframe}
\format{A data frame with 73 rows and 5 columns. Each column represents one
      bootstrap iteration.}
\usage{
sample_bootstrap
}
\description{
Bootstrap permutation vectors sample dataframe
}
\keyword{datasets}
