% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bootstrap_permutation.R
\name{sample_vector}
\alias{sample_vector}
\title{Resample a vector of values with repetition}
\usage{
sample_vector(ids)
}
\arguments{
\item{ids}{vector of the values t be resampled}
}
\value{
Same length vector with the resampled values.
}
\description{
Resample a vector of values with repetition
}
\examples{
library(lassopmm)

sample_vector(1:10)
}
