% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntheticpanel.R
\docType{package}
\name{syntheticpanel}
\alias{syntheticpanel}
\alias{syntheticpanel-package}
\title{syntheticpanel: A package for}
\description{
The package provides
}
\section{Functions}{

The foo functions ...
}

