% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_check.R
\name{var_check}
\alias{var_check}
\title{Check if the object supplied has an empty string value and retun NULL.}
\usage{
var_check(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if the object supplied has an empty string value and retun NULL.
}
