% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bootstrap_permutation.R
\name{get_bootstrap_permutation}
\alias{get_bootstrap_permutation}
\title{Generates a list of permutation vectors for bootsrapping}
\usage{
get_bootstrap_permutation(data, n_boot)
}
\arguments{
\item{data}{dataframe containing two columns \code{group} and \code{id}.}

\item{n_boot}{number of bootstrap permulations (resamplings).
If \code{n_rep = 0}, no bootstraping is performed and we have the
bootstrap sample, whcih consist of the same observations in the same
order as source data.}
}
\description{
Generates a list of permutation vectors for bootsrapping
}
\examples{
library(dplyr)
library(purrr)
library(tidyr)
library(glmnet)
library(syntheticpanel)

df <- tibble::tibble(id = 1:30, group = sort(rep(1:3, 10)))
get_bootstrap_permutation(df, 3)
}
